+++ packages/stream_chat/test/src/db/chat_persistence_client_test.dart
@@ -1,4 +1,5 @@
 import 'package:stream_chat/src/core/api/requests.dart';
+import 'package:stream_chat/src/core/api/sort_order.dart';
 import 'package:stream_chat/src/core/models/channel_model.dart';
 import 'package:stream_chat/src/core/models/channel_state.dart';
 import 'package:stream_chat/src/core/models/event.dart';
@@ -67,7 +68,7 @@ class TestPersistenceClient extends ChatPersistenceClient {
   @override
   Future<List<ChannelState>> getChannelStates(
           {Filter? filter,
-          List<SortOption<ChannelState>>? channelStateSort,
+          SortOrder<ChannelState>? channelStateSort,
           PaginationParams? paginationParams}) =>
       throw UnimplementedError();